Moving find user repository to user_repository file
